The first update of ListTreeTest.java
Hi Dr. K,

I tested the rest of methods in the ListTreeTest.java

I was wondering if it looks ok. Then I will revised it later.
I have not done the exception test yet. Just want to make sure that if I need to test all of the Exceptions.

Best,
Lulu