testMoveSubTreeKK() seems has a problem
I found the moveSubTree() method does not pass the test. I tried to add a node with key "third" to the node "second" which is a child of "root", then move the subtree rooted by "third" to "root". the result shows that "third"'s parent is still "second", and "root" is not "third"'s parent. Since this method is used in TreeGameWorld class. When I tested TreeGameWorld class, move() method is not could not pass.